[Fix]Delete button and completed button

+++ models/inventory_line.py
@@ -1,6 +1,5 @@
-from email.policy import default
-
 from odoo import fields, models, api, _
+from odoo.exceptions import UserError
 
 
 class InventoryLine(models.Model):
@@ -40,20 +39,34 @@ class InventoryLine(models.Model):
 
     #Việt
     def action_history(self):
+        self.ensure_one()
         action = {
             'name': _('History'),
             'view_mode': 'list,form',
             'res_model': 'stock.move.line',
             'views': [(self.env.ref('stock.view_move_line_tree').id, 'list'), (False, 'form')],
             'type': 'ir.actions.act_window',
-            'context': {
-                'search_default_inventory': 1,
-                'search_default_done': 1,
-                'search_default_product_id': self.product_id.id,
-            }
+            'domain': [
+                ('product_id', '=', self.product_id.id),
+                ('state', '=', 'done'), ]
+
         }
         return action
 
     def action_delete(self):
-        self.unlink()  # Xóa dòng kiểm kê hiện tại
-        return {'type': 'ir.actions.client', 'tag': 'reload'}
+        self.ensure_one()
+        product_name = self.product_id.display_name
+        if self.diff_quantity == 0:
+            raise UserError(
+                _("Unable to delete unapplied line checklist . Please apply or modify before deleting.."))
+        self.unlink()
+        return {
+            'type': 'ir.actions.client',
+            'tag': 'display_notification',
+            'params': {
+                'message': _("Dòng kiểm kê cho sản phẩm '%s' đã được xóa thành công." % product_name),
+                'type': 'success',
+                'sticky': False,
+                'next': {'type': 'ir.actions.act_window_close'},
+            }
+        }
